@@ -2,11 +2,11 @@ apply plugin: 'com.android.application'
 
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.koby.friendlocation"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
         multiDexEnabled true
@@ -37,9 +37,11 @@ dependencies {
     implementation 'com.github.GoodieBag:Pinview:v1.4'
     implementation 'com.google.firebase:firebase-storage:19.1.0'
     testImplementation 'junit:junit:4.12'
+    implementation 'androidx.preference:preference:1.1.0'
     implementation 'de.hdodenhof:circleimageview:3.0.1'
     implementation 'com.github.bumptech.glide:glide:4.9.0'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation "android.arch.lifecycle:extensions:1.1.1"
     implementation 'com.google.android.material:material:1.1.0-alpha10'
     implementation 'com.google.android.gms:play-services-location:12.0.0'
     androidTestImplementation 'androidx.test.ext:junit:1.1.0'
