@@ -41,6 +41,7 @@ public class GroupAdapter extends RecyclerView.Adapter<GroupAdapter.GroupHolder>
     @Override
     public void onBindViewHolder(@NonNull GroupHolder holder, int position) {
         holder.groupName.setText(groupList.get(position).getGroupName());
+        holder.circleLetter.setText(groupList.get(position).getGroupName().charAt(0)+"");
     }
 
     @Override
@@ -51,11 +52,13 @@ public class GroupAdapter extends RecyclerView.Adapter<GroupAdapter.GroupHolder>
     public class GroupHolder extends RecyclerView.ViewHolder{
 
         public TextView groupName;
-
+        public TextView circleLetter;
 
         public GroupHolder(@NonNull View itemView) {
             super(itemView);
             groupName = itemView.findViewById(R.id.adapter_group_name);
+            circleLetter = itemView.findViewById(R.id.circle_letter_tv);
+
             itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
